Renaming Provider Classes to Driver Classes; Deleting the old implementation for Driver support
